+++ db/migrate/20230710091311_add_deactivated_at_to_user.rb
@@ -1,0 +1,5 @@
+class AddDeactivatedAtToUser < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :deactived_at, :datetime
+  end
+end